Fixes #29: moved script into package, requirements
