+++ .env.template
@@ -11,3 +11,4 @@ NACHET_RESOURCE_GROUP=
 NACHET_WORKSPACE=
 NACHET_MODEL=
 MICROSCOPE_URL=
+METHODS=
