+++ custom_exceptions.py
@@ -8,3 +8,11 @@ class ProcessInferenceResultError(Exception):
 
 class ValidateEnvVariablesError(Exception):
     pass
+
+
+class ServerError(Exception):
+    pass
+
+
+class MicroscopeQueryError(Exception):
+    pass
